backend/requirements.txt
Flask==3.0.1
openai==1.13.3
Flask-Cors==4.0.0
python-dotenv==1.0.1